// src/types/ProductCleanupTableInfo.ts
type ProductCleanupTableInfo = {
  productId?: string;
  manufacturerId?: string;
  lastKnownSnapshotId?: string;
  productName?: string;
  productDescription?: string;
  productUrl?: string;
  imageUrl?: string;
  ownerId?: string;
  m1?: string;
  m2?: string;
  m3?: string;
  m4?: string;
  m5?: string;
  m6?: string;
  m7?: string;
  m8?: string;
  m9?: string;
  m10?: string;
};

export default ProductCleanupTableInfo;